+++ CHANGELOG.md
@@ -11,6 +11,7 @@ script to proceed.
 - Move `temboard-migratedb` as `temboard migratedb` subcommand.
 - Move `temboard-agent-register` as `temboard-agent register` subcommand.
 - New command `temboard generate-key` for generating signing key.
+- New command `temboard-agent fetch-key` to accept UI signing key.
 - New subcommands `temboard query-agent`, `temboard routes`, `temboard
   runtask`, `temboard schedule`, `temboard web`, `temboard-agent routes`,
   `temboard-agent runtask` and `temboard-agent web` for debugging.